persona-generation/requirements.txt
google-cloud-aiplatform 
tqdm
psycopg2-binary
tqdm
bcrypt